Trajactory Function

+++ titans.h
@@ -6,6 +6,6 @@ void creeVille(char ville[45][80]);
 void Affichage(int J1,int J2, char ville[45][80]);
 void refresh();
 int TailleTab(int tab[]);
-void delai(int a);
+void delai(float a);
 
 #endif
